+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/TissueType.java
@@ -1,15 +1,8 @@
 package uk.ac.bbsrc.tgac.miso.core.data;
 
 import java.io.Serializable;
-import java.util.Date;
 
-import com.eaglegenomics.simlims.core.User;
-
-public interface TissueType extends Serializable {
-
-  public long getId();
-
-  public void setId(long tissueTypeId);
+public interface TissueType extends Deletable, Identifiable, Serializable, Timestamped {
 
   public String getAlias();
 
@@ -18,22 +11,6 @@ public interface TissueType extends Serializable {
   public String getDescription();
 
   public void setDescription(String description);
-
-  public User getCreatedBy();
-
-  public void setCreatedBy(User createdBy);
-
-  public Date getCreationDate();
-
-  public void setCreationDate(Date creationDate);
-
-  public User getUpdatedBy();
-
-  public void setUpdatedBy(User updatedBy);
-
-  public Date getLastUpdated();
-
-  public void setLastUpdated(Date lastUpdated);
 
   public String getItemLabel();